+++ CHANGELOG.md
@@ -100,6 +100,7 @@ All notable changes to this project will be documented in this file.
 - Added Terminus Machina for Party and Leader Option for Experiment Purposes
 - Made the Terminate Function more function faster than before
 - Added Check Notifications
+- Added Holia Keldrasil for dungeon automation
 
 ## v5.11 Fixes